more accurate event & display init

+++ src/box.c
@@ -89,7 +89,7 @@ void box_list_contents(box *bx)
     {
       /*      if (ptr->type == WIDGET_BUTTON)
 	      printf("listing %s\n", ((button *)ptr->data)->txt); */
-      printf("size is %u\n", sizeof(*(ptr->data)));
+      printf("size is %i\n", (int)sizeof(*(ptr->data)));
       ptr = ptr->next;
     }
 }
